seperated components and implemented comparing queries to one another
