+++ src/App.tsx
@@ -1,28 +1,34 @@
 import QueryMonitor from './components/QueryMonitor';
-import TestQueryPage from './components/TestQueryPage';
+import TestQueryPage from './components/QueryPerformance/TestQueryPage';
 import AuthPage from './components/QueryMonitor/AuthPage';
 import AuthCallback from './components/QueryMonitor/AuthCallback';
 import ProtectedRoute from './components/ProtectedRoute';
 import { Route, Routes } from 'react-router-dom';
-
 
 function App() {
   return (
     <div>
       <Routes>
         {/* Auth routes */}
-        <Route path="/auth" element={<AuthPage />} />
-        <Route path="/auth/github/callback" element={<AuthCallback />} />
+        <Route path='/auth' element={<AuthPage />} />
+        <Route path='/auth/github/callback' element={<AuthCallback />} />
         {/* Protected routes */}
-        <Route path="/" element={
-          <ProtectedRoute>
-            <QueryMonitor />
-          </ProtectedRoute>
-        } />
-        <Route path='/test-query' element={
-          <ProtectedRoute>
-            <TestQueryPage />
-          </ProtectedRoute>} />
+        <Route
+          path='/'
+          element={
+            <ProtectedRoute>
+              <QueryMonitor />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path='/test-query'
+          element={
+            <ProtectedRoute>
+              <TestQueryPage />
+            </ProtectedRoute>
+          }
+        />
       </Routes>
     </div>
   );
